+++ NEWS.md
@@ -2,6 +2,7 @@
 # eatGADS 0.16.0.9000
 
 * `fac2dummies()` and `fac2dummies_complex()` for recoding a labeled factor variable to multiple labeled dummy variables
+* `equalGADS()` for comparing the complete structure and content of two `GADSdat` objects
 * `reuseMeta()` now allows only transferring value labels for values coded as missing
 
 # eatGADS 0.16.0